[SNSSQS] Write-only broker option
For SNSSQS implementation, current validation of options and use enforces SQS related options configuration even when you do not have any need to subscribe to the topic in a specific component/service using the broker. This is sticky, forcing you to maintain infrastructure and coupling to it in places where you don't need to.

Changed a bit to make it possible to instance a write-only snssqs broker.

I took the liberty of adding a new `WithCodec` option and making it mandatory. I think that it makes sense when there is that comment at the top of the constructor.